+++ src/de/xzise/xwarp/wrappers/economy/EconomyWrapper.java
@@ -3,6 +3,6 @@ package de.xzise.xwarp.wrappers.economy;
 public interface EconomyWrapper {
 
     AccountWrapper getAccount(String name);
-    String format(double price);
+    String format(int price);
     
 }
